Do not pass AWS_ credentials to boto AWS functions if thety are not set
In production (when the lambda is running on AWS) we should not pass the
AWS credential type arguments. Not passing them forces the lambda to operate
under the assumed role, and passing them means the role definition is not
correct.

Locally, when these params are set, we need to pass them explicitly.