Don't raise ReadError or ProtocolError to Rollbar
If a ReadError or ProtocolError is thrown, we want to retry ingesting the
file - we don't necessarily need to know this has happend in Rollbar. We only
need to raise errors in Rollbar if the maximum retries are exeeded, or if
another error is thrown.

This should reduce some of the noise in Rollbar for the ingester.