Get consignment reference from message in its own method
Refactor the part of the code where we extract the consignment reference from
the message. If the consignment reference doesn't exist, we should use the
tarfile name If this is an "edge" tarball it will not have a consignment
reference from TRE, so we should use whatever we can as a unique identifier for
the ingester process.