Refactor editor notification flow to new/updated judgments
Trello: https://trello.com/c/hfOotu2k

Editors were receiving occasional notifications of a new judgment, but the
attached link to the editorial UI lead to a 404 page. On investigation,
it appears two judgments have generated this email but are not in Marklogic.

In order to track down why this might be happening, we have refactored the
judgment insertion/update flow. There is a possibility the call to
`api_client.insert_judgment_xml` was failing, but not exception was being
caught and the notification of a new judgment was being sent anyway.

This refactor should hopefully catch any errors with updating/inserting a
judgment and not allow a notification to be sent if either fails.

This refactor also adds some new tests.