Refactor how we get and parse the XML judgment from the tar file
We want to reduce the number of times a judgment has to go back for re-parsing
when it appears as a "failure" in the editor UI.

As a step towards this, we want more judgments to be available in the editor UI
for editing, especially if the error in parsing is something relatively easy
to fix, for example if the neutral citation is missing.

Previously, if a judgment XML was completely missing from a tar file, or
malformed, we rejected the entire ingestion. This meant the editors were not
able to download the judgment docx in the editorial UI and potentially diagnose
the issue.

Now, we attempt to still create an entry in Marklogic, even if the judgment XML
is missing or malformed. We do this via the following logic:

- If a tarball contains the expected judgment XML and correct URI, insert it
as expected

- If a tarball contains the expected judgment XML and no URI, create a "failure"
URI in Marklogic and insert the judgment XML at that URI

- If a tarball contains malformed XML and the correct URI, create a record in
Marklogic at the correct URI and use the contents of the parser.log as the XML
document

- If a tarball contains malformed XML and no URI, create a "failure" URI in
Marklogic and use the contents of the parser.log as the XML document

- If a tarball contains no XML and no URI, create a "failure" URI in
Marklogic and use the contents of the parser.log as the XML document

If the tarball contains no XML file, log a warning but do not fail the ingestion