Bug: Catch a `FileNotFoundException` here, not a `KeyError`
`copy_file()` will throw a FileNotFoundException if an expected file is not
found in the tarfile. The KeyError is being thrown & caught inside copy_file(),
and copy_file() is retuning a FileNotFoundException.

We were previously looking to catch a KeyError in the handler, and because a
FileNotFoundException was being thrown, we weren't catching it.

This means that when a `parser.log` was not present in the tarfile, the
ingester was not saving any of the tarfile contents, including the docx. Now,
we should save the docx even if there's no parser.log present.