Refactor notification data to accept empty metadata
If there is any missing metadata in the data bag (tarfile), still try to send
a notification even if most of the info is missing.

Switching to Dict.get() means we can add some fallbacks for any of the missing
metadata. It's not ideal for editors but it's something we can iterate on.