Do not send the editors new/updated judgment notifications outside of production
The editors do not want to receive new/updated judgment notifications outside
of the production environment.

I have hung this functionality off the `ROLLBAR_ENV` environment variable as
the ingester as it stands does not have the concept of an environment. But if
we wanted to introduce an environment to the ingester, this is an easy change in
the code - we would need to add the new variable to AWS however.