Update how we perform deployments to set environment variables
Previously we didn't define any environment variables at deploy time;
instead we set them manually. A configuration change when we moved from
Python 3.9 to 3.11 caused SAM to redefine the infrastructure, and this
in turn caused a all existing environment variables to be wiped,
bringing down the staging ingester.

This PR moves to defining the environment variables in the
`template.yml` file which is used by SAM. The exact values are passed in
at deploy time by GitHub Actions as part of configuring the
infrastructure.

Variables are stored in GitHub environments.