Add new property to identify ingested document type based on XML root node
When a document is ingested, as currently deployed we rely on the presence of press-summary in the URI to hint at the type. As we move to abstract identifiers this falls apart, and instead we must rely on the content of the document itself to provide type hinting.

This adds a new property to the Ingest class to reflect the type of document being ingested, which can then be used in future to make sure we do things like correctly assign identifier types.