Swap tarfile handling in Ingest to use injected dependency
Previously the `Ingest` class was repeatedly opening and closing the same tarfile, which we believe is contributing to higher than expected memory usage.

Change the behaviour so instead of the `Ingest` class handling opening and closing, the lambda runner takes on the responsibility of downloading the `.tar.gz` from S3, opening it, and passing the open `TarFile` instance back into `Ingest`. This means we also need to pass in the intended destination filename for the re-uploaded `.tar.gz` file, but this is probably closer to correct behaviour anyway.

As a side-effect, we can remove some mocking from tests since we can directly inject the dependencies.

The `Ingest` class still re-opens the `.tar.gz` file for copying to S3, but this shouldn't be an issue with multiple open file pointers as they are all in read-only mode.